requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
WTForms==3.1.2
markupsafe==3.0.2
Flask-Migrate==4.0.5
Werkzeug==3.0.1
email-validator==2.1.1
pytest==8.0.2
python-dotenv==1.0.1 